backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
chromadb
ollama
numpy
pandas
scikit-learn
joblib
sentence-transformers
matplotlib
jsonlines